+++ step1/index.html
@@ -25,6 +25,6 @@
   </head>
   <body>
     <main id="app"></main>
-    <script src="main.js" defer></script>
+    <script type = "module" src="app.js" defer></script>
   </body>
 </html>
