+++ step1/state.js
@@ -1,0 +1,8 @@
+export const state = {
+    todoItems : [
+        {id: 1, content: '첫번째 아이템', isComplete: false, createdAt:Date.now()},
+        {id: 2, content: '두번째 아이템', isComplete: false, createdAt:Date.now()},
+        {id: 3, content: '세번째 아이템', isComplete: false, createdAt:Date.now()},
+    ],
+    selectedItem: -1,
+}
