+++ step2/src/bar.js
@@ -1,0 +1,3 @@
+export default function bar() {
+    //
+  }
