+++ step2/src/main.js
@@ -1,0 +1,3 @@
+import bar from './bar.js';
+
+bar();
